check if zip file exits

+++ capture.py
@@ -1,7 +1,8 @@
 
-import cv2
+import os
 import requests
 from tqdm import tqdm
+import cv2
 
 # creating an object, zero for external camera
 cap = cv2.VideoCapture(0)
@@ -15,11 +16,13 @@ else:
 
 url = 'https://github.com/davisking/dlib-models/raw/master/shape_predictor_68_face_landmarks.dat.bz2'
 local_filename = url.split('/')[-1]
-response = requests.get(url, stream=True)
-length = response.headers.get('content-length', 0)
-with tqdm.wrapattr(open(local_filename, 'wb'), 'write', miniters=1, desc='Downloading', total=int(length)) as fout:
-    for chunk in response.iter_content(chunk_size=4096):
-        fout.write(chunk)
+
+if not os.path.exists(local_filename):
+    response = requests.get(url, stream=True)
+    length = response.headers.get('content-length', 0)
+    with tqdm.wrapattr(open(local_filename, 'wb'), 'write', miniters=1, desc='Downloading', total=int(length)) as fout:
+        for chunk in response.iter_content(chunk_size=4096):
+            fout.write(chunk)
 
 # # for photo capture
 # img = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
